+++ pkg/providers/iam/testdata/TestResourceAPIClient/create_all_accessible_apis_false_apis_null.tf
@@ -1,0 +1,26 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_iam_api_client" "test" {
+  authorized_users    = ["mw+2"]
+  client_type         = "CLIENT"
+  client_name         = "mw+2_1"
+  notification_emails = ["mw+2@example.com"]
+  client_description  = "Test API Client"
+  lock                = false
+  credential          = {}
+  group_access = {
+    clone_authorized_user_groups = false
+    groups = [
+      {
+        group_id = 123
+        role_id  = 340
+      }
+    ]
+  }
+  api_access = {
+    all_accessible_apis = false
+    apis                = null
+  }
+}